feat: day 9 + restructuring

+++ IDays.cs
@@ -2,5 +2,5 @@ namespace AOC24;
 
 public interface IDays
 {
-    public int Solve();
+    public long Solve();
 }
